Initial pass at restructuring component
This relocates the various progressive disclosure panel components into different subdirectories as an initial step to bringing this component into inline with newer components being added to Gutenberg.
