+++ packages/components/src/index.js
@@ -100,8 +100,10 @@ export { default as PanelHeader } from './panel/header';
 export { default as PanelRow } from './panel/row';
 export { default as Placeholder } from './placeholder';
 export { default as Popover } from './popover';
-export { default as __experimentalProgressiveDisclosurePanel } from './progressive-disclosure-panel';
-export { default as __experimentalProgressiveDisclosurePanelItem } from './progressive-disclosure-panel/item';
+export {
+	ProgressiveDisclosurePanel as __experimentalProgressiveDisclosurePanel,
+	ProgressiveDisclosurePanelItem as __experimentalProgressiveDisclosurePanelItem,
+} from './progressive-disclosure-panel';
 export { default as QueryControls } from './query-controls';
 export { default as __experimentalRadio } from './radio';
 export { default as __experimentalRadioGroup } from './radio-group';
