+++ packages/components/src/progressive-disclosure-panel/progressive-disclosure-panel-title/component.js
@@ -7,10 +7,10 @@ import { __ } from '@wordpress/i18n';
 /**
  * Internal dependencies
  */
-import { usePanelContext } from './';
-import MenuGroup from '../menu-group';
-import MenuItem from '../menu-item';
-import DropdownMenu from '../dropdown-menu';
+import { usePanelContext } from '../progressive-disclosure-panel';
+import MenuGroup from '../../menu-group';
+import MenuItem from '../../menu-item';
+import DropdownMenu from '../../dropdown-menu';
 
 const ProgressiveDisclosurePanelTitle = ( props ) => {
 	const { menuLabel, resetAll, title, toggleChild } = props;
