+++ packages/components/src/progressive-disclosure-panel/progressive-disclosure-panel-title/index.js
@@ -1,0 +1,1 @@
+export { default } from './component';
