+++ packages/components/src/progressive-disclosure-panel/progressive-disclosure-panel/index.js
@@ -1,0 +1,1 @@
+export { default, usePanelContext } from './component';
